Unhide ToPartialReflect and ToReflect
Also moved these traits to a new `cast` module
